Subscribers dashboard: Forms-style layout polish (#48365)
Restructure the page chrome to mirror Jetpack Forms' dashboard layout:

- Drop the SubscribersTotals header line and the AdminPage wrapper.
  Replace with @wordpress/admin-ui's Page rendered INSIDE the stage
  surface, so the title + actions live as part of the stage card
  instead of a global header above the layout.
- Stage and inspector are now equal flex siblings in a 1:1 copy of
  Forms' .jp-forms-layout__surface CSS — mobile-first fixed-fullscreen
  with z-index layering, switching to inline-flex with border-radius
  cards on min-width: 782px (Forms' break-medium).
- Dark canvas: body bg matches the wp-admin sidebar (#1d2327, fresh
  scheme); light #fcfcfc surface cards pop against it. Hide #wpfooter,
  zero #wpcontent's padding-left and #wpbody-content's padding-bottom
  — same chrome overrides Forms uses.
- Side panel via plain CSS + a viewport-driven Modal fallback on
  mobile (useViewportMatch( 'medium', '<' )), separating the
  layout-agnostic SubscriberDetailContent from the inline panel /
  Modal wrapper. Detail data shape includes the country object
  ({ code, name }) it actually has, with country.name rendered
  instead of the object (fixes the React "objects are not valid as
  a React child" crash).
- Belt-and-suspenders: index.tsx tags <body> with
  jetpack-subscribers-dashboard-page so the chrome overrides match
  regardless of which WP-generated body class the menu lands on.
- New deps: @automattic/jetpack-components, @automattic/jetpack-script-data,
  @wordpress/admin-ui, @wordpress/compose.

The dark canvas hard-codes #1d2327 (default "fresh" admin color
scheme); other admin schemes (light/modern/coffee) will desync. We
can swap to a JS-detected sidebar color later if needed.
