+++ projects/packages/subscribers-dashboard/changelog/phase-6-ui-polish-side-panel
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Subscribers dashboard: Forms-style flex layout (stage + inspector as sibling surfaces), Page header inside the stage card, dark canvas matching the wp-admin sidebar.
