+++ projects/packages/subscribers-dashboard/package.json
@@ -36,10 +36,13 @@
 	},
 	"dependencies": {
 		"@automattic/babel-plugin-replace-textdomain": "workspace:*",
+		"@automattic/jetpack-components": "workspace:*",
 		"@automattic/jetpack-script-data": "workspace:*",
 		"@tanstack/react-query": "5.90.8",
+		"@wordpress/admin-ui": "1.12.0",
 		"@wordpress/api-fetch": "7.44.0",
 		"@wordpress/components": "32.6.0",
+		"@wordpress/compose": "7.44.0",
 		"@wordpress/data": "10.44.0",
 		"@wordpress/dataviews": "14.1.0",
 		"@wordpress/date": "5.44.0",
